pyproject.toml
[project]
name = "querycraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "faker>=37.11.0",
    "langchain>=0.3.27,<1.0.0",
    "langchain-ollama>=0.3.10",
    "langfuse>=3.7.0",
    "langgraph>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.2,<2.12.3",
    "sqlalchemy>=2.0.44",
]
